+++ src/log_manip.cpp
@@ -75,7 +75,7 @@ void LogManipulator::return_statistics_table () {
         (*poos_) << std::right << std::setw(colWidth) << stat_name << " ";
     }
     (*poos_) << std::endl;
-    for (int i = 1; i < num_cols_; i++) {
+    for (unsigned long i = 1; i < static_cast<unsigned long>(num_cols_); i++) {
         calculate_summary_statistics (parm_samples_[i], mean, variance, median,
             ESS, ACT, nsamples, step_size);
         int diff = longest_label - static_cast<int>(parm_names_[i].size());
@@ -134,15 +134,15 @@ void LogManipulator::calculate_summary_statistics (std::vector<double>& vals,
     // anyway, keep it like this for now to make sure identical values are computed
     
     int MAXIMUM_LAG = 2000;
-    int max_lag = std::min(n_samples - 1, MAXIMUM_LAG);
+    unsigned long max_lag = static_cast<unsigned long>(std::min(n_samples - 1, MAXIMUM_LAG));
     std::vector<double> gamma_stat(max_lag);
     
     double var_stat = 0.0;
     double del1 = 0.0;
     double del2 = 0.0;
 
-    for (int i = 0; i < max_lag; i++) {
-        for (int j = 0; j < n_samples - i; j++) {
+    for (unsigned long i = 0; i < max_lag; i++) {
+        for (unsigned long j = 0; j < n_samples - i; j++) {
             del1 = vals[j] - mean;
             del2 = vals[j + i] - mean;
             gamma_stat[i] += (del1 * del2);
@@ -188,7 +188,7 @@ void LogManipulator::collect_parameter_samples () {
         bool first_entry = true; // use for initialization
         std::vector <double> terp; // a vector to reuse throughout
         ntotal_samples_ = 0;
-        for (int i = 0; i < num_files_; i++) {
+        for (unsigned long i = 0; i < static_cast<unsigned long>(num_files_); i++) {
             std::string curfile = files_[i];
             infilestr_.open(curfile.c_str());
             std::string line;
@@ -206,8 +206,8 @@ void LogManipulator::collect_parameter_samples () {
                         parm_names_ = header;
                         if (first_entry) {
                             // let us initialize the sample collector
-                            int n_expected = 10000000; // purposely an overestimate to avoid reallocation
-                            for (int j = 0; j < num_cols_; j++) {
+                            unsigned long n_expected = 10000000; // purposely an overestimate to avoid reallocation
+                            for (unsigned long j = 0; j < static_cast<unsigned long>(num_cols_); j++) {
                                 parm_samples_.push_back(terp);
                                 parm_samples_[j].reserve(n_expected);
                             }
@@ -258,7 +258,7 @@ void LogManipulator::collect_parameter_samples () {
             infilestr_.close();
         }
         if (verbose_) {
-            for (int i = 0; i < num_files_; i++) {
+            for (unsigned long i = 0; i < static_cast<unsigned long>(num_files_); i++) {
                 std::cout << files_[i] << ": " << indiv_sample_totals_[i]
                     << " samples retained (from original " << indiv_raw_counts_[i]
                     << " samples) for " << (num_cols_ - 1) << " variables." << std::endl;
@@ -280,7 +280,7 @@ void LogManipulator::store_sample (std::string const& line) {
     // convert to double. note that this include the first entry (state) which is an int
     std::vector<double> converted = string_v_to_double_v(sample);
     // finally, store these puppies in the initialized container parm_samples_
-    for (int i = 0; i < num_cols_; i++) {
+    for (unsigned long i = 0; i < static_cast<unsigned long>(num_cols_); i++) {
         parm_samples_[i].push_back(converted[i]);
     }
 }
@@ -300,7 +300,7 @@ void LogManipulator::count_parameter_samples () {
     num_cols_ = 0;
     if (!files_.empty()) {
         for (int i = 0; i < num_files_; i++) {
-            std::string curfile = files_[i];
+            std::string curfile = files_[static_cast<unsigned long>(i)];
             infilestr_.open(curfile.c_str());
             std::string line;
             bool first_line = true;
@@ -353,7 +353,7 @@ void LogManipulator::count_parameter_samples () {
 void LogManipulator::count_tree_samples () {
     if (!files_.empty()) {
         for (int i = 0; i < num_files_; i++) {
-            std::string curfile = files_[i];
+            std::string curfile = files_[static_cast<unsigned long>(i)];
             infilestr_.open(curfile.c_str());
             std::string line;
             int num_samps = 0;
@@ -386,7 +386,7 @@ void LogManipulator::count_tree_samples () {
 
 void LogManipulator::get_sample_counts () {
     if (logtype_ == "parameter") {
-        for (int i = 0; i < num_files_; i++) {
+        for (unsigned long i = 0; i < static_cast<unsigned long>(num_files_); i++) {
             (*poos_) << files_[i] << ": " << indiv_sample_totals_[i] << " samples of "
                 << (num_cols_ - 1) << " variables." << std::endl;
         }
@@ -395,7 +395,7 @@ void LogManipulator::get_sample_counts () {
                 << (num_cols_ - 1) << " variables across " << num_files_ << " files." << std::endl;
         }
     } else {
-        for (int i = 0; i < num_files_; i++) {
+        for (unsigned long i = 0; i < static_cast<unsigned long>(num_files_); i++) {
             (*poos_) << files_[i] << ": " << indiv_sample_totals_[i] << " trees." << std::endl;
         }
         if (num_files_ > 1) {
@@ -426,7 +426,7 @@ void LogManipulator::get_column_names () {
                 }
             }
             for (int i = 0; i < num_cols_; i++) {
-                (*poos_) << i+1 << ". " << parm_names_[i] << std::endl;
+                (*poos_) << i+1 << ". " << parm_names_[static_cast<unsigned long>(i)] << std::endl;
             }
         } else {
         // multiple files. make sure number/name/order of columns is identical
@@ -458,7 +458,7 @@ void LogManipulator::delete_columns (const std::vector<int>& col_ids) {
                 num_cols_ = curpars;
                 parm_names_ = header;
                 
-                cols_to_retain.resize(num_cols_);
+                cols_to_retain.resize(static_cast<unsigned long>(num_cols_));
                 
                 // check that right end of col_ids is valid (0 already checked upstream)
                 // vector has been sorted
@@ -477,7 +477,7 @@ void LogManipulator::delete_columns (const std::vector<int>& col_ids) {
                 
                 num_cols_retain_ = static_cast<int>(cols_to_retain.size());
                 for (int i = 0; i < num_cols_retain_; i++) {
-                    (*poos_) << header[cols_to_retain[i]];
+                    (*poos_) << header[cols_to_retain[static_cast<unsigned long>(i)]];
                     if (i < (num_cols_retain_ - 1)) {
                         (*poos_) << "\t";
                     }
@@ -488,7 +488,7 @@ void LogManipulator::delete_columns (const std::vector<int>& col_ids) {
             } else {
                 std::vector<std::string> samp = tokenize(line);
                 for (int i = 0; i < num_cols_retain_; i++) {
-                    (*poos_) << samp[cols_to_retain[i]];
+                    (*poos_) << samp[cols_to_retain[static_cast<unsigned long>(i)]];
                     if (i < (num_cols_retain_ - 1)) {
                         (*poos_) << "\t";
                     }
@@ -539,7 +539,7 @@ void LogManipulator::retain_columns (const std::vector<int>& col_ids) {
                 
                 parm_names_ = header;
                 for (int i = 0; i < num_cols_retain_; i++) {
-                    (*poos_) << header[cols_to_retain[i]];
+                    (*poos_) << header[cols_to_retain[static_cast<unsigned long>(i)]];
                     if (i < (num_cols_retain_ - 1)) {
                         (*poos_) << "\t";
                     }
@@ -550,7 +550,7 @@ void LogManipulator::retain_columns (const std::vector<int>& col_ids) {
             } else {
                 std::vector<std::string> samp = tokenize(line);
                 for (int i = 0; i < num_cols_retain_; i++) {
-                    (*poos_) << samp[cols_to_retain[i]];
+                    (*poos_) << samp[cols_to_retain[static_cast<unsigned long>(i)]];
                     if (i < (num_cols_retain_ - 1)) {
                         (*poos_) << "\t";
                     }
@@ -572,7 +572,7 @@ void LogManipulator::sample_parameters () {
     if (!files_.empty()) { // hrm this should not be necessary...
         ntotal_samples_ = 0;
         for (int i = 0; i < num_files_; i++) {
-            std::string curfile = files_[i];
+            std::string curfile = files_[static_cast<unsigned long>(i)];
             infilestr_.open(curfile.c_str());
             std::string line;
             bool first_line = true;
@@ -588,7 +588,7 @@ void LogManipulator::sample_parameters () {
                         num_cols_ = curpars;
                         parm_names_ = header;
                         for (int j = 0; j < num_cols_; j++) {
-                            (*poos_) << parm_names_[j];
+                            (*poos_) << parm_names_[static_cast<unsigned long>(j)];
                             if (j < (num_cols_ - 1)) {
                                 (*poos_) << "\t";
                             }
@@ -639,7 +639,7 @@ void LogManipulator::sample_parameters () {
             infilestr_.close();
         }
         if (verbose_) {
-            for (int i = 0; i < num_files_; i++) {
+            for (unsigned long i = 0; i < static_cast<unsigned long>(num_files_); i++) {
                 std::cout << files_[i] << ": " << indiv_sample_totals_[i]
                     << " samples retained (from original " << indiv_raw_counts_[i]
                     << " samples) for " << (num_cols_ - 1) << " variables." << std::endl;
@@ -655,7 +655,7 @@ void LogManipulator::sample_trees () {
     if (!files_.empty()) {
         ntotal_samples_ = 0;
         for (int i = 0; i < num_files_; i++) {
-            std::string curfile = files_[i];
+            std::string curfile = files_[static_cast<unsigned long>(i)];
             infilestr_.open(curfile.c_str());
             std::string line;
             int tree_counter = 0; // this is the raw number of tree lines in a file
@@ -719,7 +719,7 @@ void LogManipulator::sample_trees () {
         (*poos_) << "End;" << std::endl;
         
         if (verbose_) {
-            for (int i = 0; i < num_files_; i++) {
+            for (unsigned long i = 0; i < static_cast<unsigned long>(num_files_); i++) {
                 std::cerr << files_[i] << ": " << indiv_sample_totals_[i]
                     << " tree samples retained (from original " << indiv_raw_counts_[i]
                     << " samples)." << std::endl;
@@ -747,7 +747,7 @@ void LogManipulator::write_reformatted_sample (std::string& sample, int& sample_
     if (logtype_ == "parameter") {
         (*poos_) << sample_num;
         for (int i = 1; i < num_cols_; i++) {
-            (*poos_) << "\t" << terp[i];
+            (*poos_) << "\t" << terp[static_cast<unsigned long>(i)];
         }
         (*poos_) << std::endl;
     } else {
