+++ src/nj.h
@@ -25,13 +25,13 @@ private:
     void CalcQ (const int& NumbOfSequences, std::vector< std::vector<double> >& OriginalMatrix, 
         std::vector< std::vector<double> >& ConvertedMatrix, std::vector< std::vector<double> >& LengthMatrix);
     void FetchLengths (const int& NumbOfSequences, const std::vector< std::vector<double> >& NewMatrix,
-        std::vector< std::vector<double> >& LengthMatrix, const int& mini1, const int& mini2,
-        double& brlength1, double& brlength2);
+        std::vector< std::vector<double> >& LengthMatrix, const unsigned long& mini1,
+        const unsigned long& mini2, double& brlength1, double& brlength2);
     void Tree_Update (std::string& newname, std::vector<std::string>& names,
-        int& NumbOfSequences, std::vector< std::vector<double> >& NewMatrix, int& mini1, int& mini2,
-        double& brlength1, double& brlength2);
+        int& NumbOfSequences, std::vector< std::vector<double> >& NewMatrix, unsigned long& mini1,
+        unsigned long& mini2, double& brlength1, double& brlength2);
     void Choose_Smallest (int& NumbOfSequences, const std::vector< std::vector<double> >& Matrix,
-        int& mini1, int& mini2);
+        unsigned long& mini1, unsigned long& mini2);
     
 public:
     NJOI (std::istream* pios, int& threads);
